ipfs-proxy/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
ez-setup==0.9
Flask==2.2.2
idna==3.4
ipfs-api==0.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
pip-autoremove==0.10.0
PyPDF2==3.0.0
qrcode==7.3.1
requests==2.28.1
six==1.16.0
urllib3==1.26.13
waitress==2.1.2
Werkzeug==2.2.2
